requirements.txt
Django==1.6.7
Fabric==1.10.0
Pillow==2.5.3
South==1.0
argparse==1.2.1
django-suit==0.2.11
paramiko==1.15.1
wsgiref==0.1.2
